+++ isitfit/cli/core.py
@@ -15,6 +15,7 @@ import click
 from ..utils import display_footer
 from .. import isitfit_version
 
+
 # With atexit, this message is being displayed even in case of early return or errors.
 # Changing to try/finally in the __main__ below
 #import atexit
@@ -25,11 +26,12 @@ from .. import isitfit_version
 
 @click.group(invoke_without_command=True)
 @click.option('--debug', is_flag=True, help='Display more details to help with debugging')
-@click.option('--optimize', is_flag=True, help='DEPRECATED: get cost optimization recommendations')
+@click.option('--optimize', is_flag=True, help='DEPRECATED: use "isitfit cost optimize" instead')
+@click.option('--version', is_flag=True, help='DEPRECATED: use "isitfit version" instead')
 @click.option('--share-email', multiple=True, help='Share result to email address')
+@click.option('--skip-check-update', is_flag=True, help='Skip step for checking for update')
 @click.pass_context
-def cli_core(ctx, debug, optimize, share_email):
-
+def cli_core(ctx, debug, optimize, version, share_email, skip_check_update):
     logLevel = logging.DEBUG if debug else logging.INFO
     ch = logging.StreamHandler()
     ch.setLevel(logLevel)
@@ -40,13 +42,27 @@ def cli_core(ctx, debug, optimize, share_email):
       logger.debug("Enabled debug level")
       logger.debug("-------------------")
 
+    # if a command is invoked, eg `isitfit tags`, do not proceed
+    if ctx.invoked_subcommand is None:
+        # if still used without subcommands, notify user of new usage
+        #from .cost import analyze as cost_analyze, optimize as cost_optimize
+        #if optimize:
+        #  ctx.invoke(cost_optimize, filter_tags=filter_tags, n=n)
+        #else:
+        #  ctx.invoke(cost_analyze, filter_tags=filter_tags)
+        if optimize:
+          click.secho("As of version 0.11, please use `isitfit cost optimize` instead of `isitfit --optimize`.", fg='red')
+        else:
+          click.secho("As of version 0.11, please use `isitfit cost analyze` instead of `isitfit` to calculate the cost-weighted utilization.", fg='red')
+
+        # just return non-0 code
+        import sys
+        sys.exit(1)
+
     # check if current version is out-of-date
-    from ..utils import prompt_upgrade
-    is_outdated = prompt_upgrade('isitfit', isitfit_version)
-    if is_outdated:
-      # Give the user some time to read the message and possibly update
-      import time
-      time.sleep(3)
+    if not skip_check_update:
+      from ..utils import prompt_upgrade
+      prompt_upgrade('isitfit', isitfit_version)
 
     # make sure that context is a dict
     ctx.ensure_object(dict)
@@ -55,30 +71,17 @@ def cli_core(ctx, debug, optimize, share_email):
     if share_email is not None:
       ctx.obj['share_email'] = share_email
 
+    # Important that this be "after" the check for update
+    if version:
+        # ctx.invoke(cli_version)
+        click.secho("As of version 0.11, please use `isitfit version` instead of `isitfit --version`.", fg='red')
+        import sys
+        sys.exit(1)
+
     # After adding the separate command for "cost" (i.e. `isitfit cost analyze`)
     # putting a note here to notify user of new usage
     # Ideally, this code would be deprecated though
 
-    # if a command is invoked, eg `isitfit tags`, do not proceed
-    if ctx.invoked_subcommand is not None:
-      return
-
-    # if still used without subcommands, notify user of new usage
-    #from .cost import analyze as cost_analyze, optimize as cost_optimize
-    #if optimize:
-    #  ctx.invoke(cost_optimize, filter_tags=filter_tags, n=n)
-    #else:
-    #  ctx.invoke(cost_analyze, filter_tags=filter_tags)
-    if optimize:
-      click.secho("Deprecation note as of version 0.11:", fg='red')
-      click.secho("  Please use `isitfit cost optimize` instead of `isitfit --optimize`.", fg='red')
-    else:
-      click.secho("Deprecation note as of version 0.11:", fg='red')
-      click.secho("  Please use `isitfit cost analyze` instead of `isitfit` to calculate the cost-weighted utilization.", fg='red')
-
-    # just return non-0 code
-    import sys
-    sys.exit(1)
 
 
 from .tags import tags as cli_tags